refactor(shuttle): the vocabulary every verb is written in is a module of its own
`verbs.ts` held four unrelated things: the dispatch, the vocabulary every verb
is written in, `cd`'s adopt-a-place machine, and the verb bodies. Only the
second is what a new verb reaches for, and the writing verbs had to be written
inside that file because the vocabulary was not importable.

So `vocabulary.ts`, and it holds what a verb reaches for before it has done
anything of its own: the types (`Outcome`, `Interruption`, `Refusal`,
`Shuttle`, `VerbDeps`, `VerbLine`, `Verb`); `guarded` / `stopped` / `Ran`;
`refuse`; `container`; the operand-to-place path (`aimed`, `reading`,
`connectedSpace`, and the `%n` lookup `rowFor` that path calls); the config a
verb reaches a piece through (`pieceConfigOf`, `pieceConfigAt`); and the page
path (`paged`, `screenFit`, `measured`).

This is a move and nothing else. Every declaration is relocated verbatim. The
only edit to any of them is the `export` keyword on the sixteen a caller
outside the module now names, and the two signatures `deno fmt` rewrapped once
that keyword pushed them past the width; the rest of the diff is the import
wiring, the two module doc comments, and the module list in the README. No
signature, no behavior and no assertion changes, which is what makes the suite
the whole proof: every case passes unchanged.

What stays is what nothing else reaches for. `cd`'s settle machine — `landing`,
`settlePiece`, `resolvedPiece`, `confirmedKey`, `alreadyStoodAt`, `noSuchKey`,
`Bearing` — and no other verb adopts a place. Likewise `read` / `written`
(`get`'s), `resolveTarget` / `addressIn` (`wish`'s), the writes' and the calls'
own helpers, and `wrongOperandCount` / `tooMany` / `notAVerb` / `listed`, which
are the dispatch's own.

Closes #7052.
