+++ packages/cli/test/shuttle-prompt.test.ts
@@ -31,7 +31,7 @@ import { CurrentPlace } from "../lib/shuttle/place.ts";
 import { ShuttleSession } from "../lib/shuttle/session.ts";
 import { moved } from "./shuttle-place-helpers.ts";
 import { type PromptTerminal, runPrompt } from "../lib/shuttle/prompt.ts";
-import type { Shuttle, VerbDeps } from "../lib/shuttle/verbs.ts";
+import type { Shuttle, VerbDeps } from "../lib/shuttle/vocabulary.ts";
 import type { Key } from "../lib/view/keys.ts";
 
 const SPACE = "did:key:z6MkConnectedSpace" as MemorySpace;
